requirements.txt
matplotlib
numpy
pandas
pyyaml
google-generativeai